requirements.txt
Bootstrap_Flask==2.4.0
Requests==2.31.0
WTForms==3.1.2
Flask_WTF==1.2.1
Werkzeug==3.0.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29

python-dotenv~=1.0.1